--- a/crates/common/src/lib.rs
+++ b/crates/common/src/lib.rs
@@ -1,9 +1,9 @@
 pub mod result;
-pub mod slice;
+// pub mod slice;
 
 pub use holochain_serialized_bytes::prelude::*;
 pub use result::*;
-pub use slice::*;
+// pub use slice::*;
 
 /// something like usize for wasm
 /// wasm has a memory limit of 4GB so offsets and lengths fit in u32
@@ -19,4 +19,10 @@ pub use slice::*;
 pub type WasmSize = u32;
 
 pub type Len = WasmSize;
+
 pub type GuestPtr = WasmSize;
+
+pub fn allocation_length(len: Len) -> Len {
+    // the first bytes are the length of the allocation so we can load from it directly
+    len + std::mem::size_of::<Len>() as Len
+}