--- a/crates/common/src/result.rs
+++ b/crates/common/src/result.rs
@@ -2,7 +2,7 @@ use holochain_serialized_bytes::prelude::*;
 use thiserror::Error;
 
 /// Enum of all possible ERROR codes that a Zome API Function could return.
-#[derive(Clone, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, Serialize, Deserialize, Error)]
+#[derive(Clone, Debug, PartialEq, Eq, Hash, PartialOrd, Ord, Serialize, Deserialize, SerializedBytes, Error)]
 #[rustfmt::skip]
 pub enum WasmError {
     /// while converting pointers and lengths between u64 and i64 across the host/guest
@@ -15,7 +15,7 @@ pub enum WasmError {
     /// correctly, which should be impossible for well behaved serialization
     SerializedBytes(SerializedBytesError),
     /// something went wrong while writing or reading bytes to/from wasm memory
-    /// this means something like "reading 16 bytes did not produce 2x u64 ints"
+    /// this means something like "reading 16 bytes did not produce 2x WasmSize ints"
     /// or maybe even "failed to write a byte to some pre-allocated wasm memory"
     /// whatever this is it is very bad and probably not recoverable
     Memory,
@@ -58,13 +58,11 @@ impl std::fmt::Display for WasmError {
     }
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, SerializedBytes)]
 pub enum WasmResult {
     Ok(SerializedBytes),
     Err(WasmError),
 }
-
-holochain_serial!(WasmResult, WasmError);
 
 #[cfg(test)]
 pub mod tests {
@@ -73,10 +71,8 @@ pub mod tests {
 
     #[test]
     fn wasm_result_serialized_bytes_round_trip() {
-        #[derive(Clone, PartialEq, Debug, Serialize, Deserialize)]
+        #[derive(Clone, PartialEq, Debug, Serialize, Deserialize, SerializedBytes)]
         struct Foo(String);
-
-        holochain_serial!(Foo);
 
         let foo = Foo(String::from("bar"));