… Cargo.lock
… Cargo.toml
… README.md
… crates/common/Cargo.toml
… crates/guest/Cargo.lock
… crates/guest/Cargo.toml
… crates/host/Cargo.toml
… crates/host/src/guest.rs
… crates/host/src/import.rs
… crates/host/src/instantiate.rs
… test.sh
… test/Cargo.lock
… test/Cargo.toml
… test/benches/bench.rs
… test/build.rs
… test/common/Cargo.toml
… test/src/import.rs
… test/src/load_wasm.rs
… test/src/test.rs
… test/src/wasms.rs
… test/test_wasm/Cargo.lock
… test/test_wasm/Cargo.toml
… test/test_wasm/src/wasm.rs
… test/wasm_empty/Cargo.lock
… test/wasm_io/Cargo.lock
… test/wasm_io/Cargo.toml
… test/wasm_io/src/wasm.rs
… test/wasm_memory/Cargo.lock
… test/wasm_memory/Cargo.toml
… test/wasm_memory/src/wasm.rs