this changes the host/guest data negotiation from a system where the host and guest ask each other for the data at a specific memory location (requires trust between the host/guest) to a system where the guest can only interact with it's own memory locations and the host/guest co-ordinate via. length-prefixed data allocations, which means a malicious guest can't steal or corrupt host data (although it could temporarily leak a bit of host data for the lifetime of the host if it calls a host function and leaves the result hanging)

so basically, if a host has bytes `[ 1_u8, 2_u8, 3_u8 ]` to write into the guest, the guest is asked to allocate `3 + 4` bytes (because 4 bytes is the length of a `u32`) and then the host writes then length and then the bytes at the guest pointer.

so the bytes written are `[ 3_u8, 0_u8, 0_u8, 0_u8, 1_u8, 2_u8, 3_u8 ]`

this means that the host/guest allocators don't have to agree on anything other than:

- byte order for a length is little endian (required by wasm)
- lengths are u32 (required by wasm)
- u32 is 4 u8 bytes long (required by wasm)

and they can pass data and (de)serialize round trip `SerializedBytes` through `GuestPtr` values alone

__note a minor breaking change, `RemotePtr` is renamed to `GuestPtr` and is a `u32` instead of a `u64` now__

read the readme for more info

it also benchmarks and cleans things up that were slow (serialization and wee_alloc issues)

- remove the guest crate from the core workspace so that it only exists inside wasm guests
- update README.md
- deleted allocation.rs (all the AllocationPtr and Allocation stuff)
- change `u64` `RemotePtr` to `u32` `GuestPtr`
- add benchmarks
- no longer recommend wee_alloc as it is 30x slower to round trip serialized data than the default allocator for some reason
- use length prefixed allocations to move data between host/guest
- use serde_bytes as it handles binary data about 50x faster

indicative benchmarks:

```
[nix-shell:~/holochain-wasmer]$ ./bench.sh 
   Compiling test v0.0.31 (/home/thedavidmeister/holochain-wasmer/test)
   Compiling holochain_wasmer_common v0.0.31 (/home/thedavidmeister/holochain-wasmer/crates/common)
   Compiling holochain_wasmer_host v0.0.31 (/home/thedavidmeister/holochain-wasmer/crates/host)
   Compiling test_common v0.0.31 (/home/thedavidmeister/holochain-wasmer/test/common)
    Finished bench [optimized] target(s) in 15.71s
     Running /home/thedavidmeister/holochain-wasmer/target/release/deps/test-edba6e285dc0a486

running 8 tests
test tests::bytes_round_trip ... ignored
test tests::native_struct_test ... ignored
test tests::native_test ... ignored
test tests::process_string_test ... ignored
test tests::ret_test ... ignored
test tests::smoke_module ... ignored
test tests::stacked_test ... ignored
test tests::try_result_test ... ignored

test result: ok. 0 passed; 0 failed; 8 ignored; 0 measured; 0 filtered out

     Running /home/thedavidmeister/holochain-wasmer/target/release/deps/bench-b08a5b016c0f09ad
wasm_instance/wasm_instance/empty                                                                             
                        time:   [42.687 us 42.878 us 43.139 us]
                        change: [-7.7194% -6.9846% -6.2757%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 12 outliers among 100 measurements (12.00%)
  4 (4.00%) high mild
  8 (8.00%) high severe
wasm_instance/wasm_instance/io                                                                             
                        time:   [41.824 us 41.934 us 42.067 us]
                        change: [-8.1558% -7.8460% -7.5693%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 6 outliers among 100 measurements (6.00%)
  1 (1.00%) high mild
  5 (5.00%) high severe
wasm_instance/wasm_instance/test                                                                             
                        time:   [43.703 us 44.117 us 44.652 us]
                        change: [-4.7849% -3.1871% -1.8707%] (p = 0.00 < 0.05)
                        Performance has improved.
Found 5 outliers among 100 measurements (5.00%)
  2 (2.00%) high mild
  3 (3.00%) high severe

Benchmarking wasm_call/io string_input_args_echo_ret/0: Collecting 10 samples in estimated 5.0002 s (929k iterations                                                                                                                    wasm_call/io string_input_args_echo_ret/0                        
                        time:   [5.4191 us 5.6588 us 5.8846 us]
                        thrpt:  [0.0000   B/s 0.0000   B/s 0.0000   B/s]
                 change:
                        time:   [+5.3959% +9.3739% +12.514%] (p = 0.00 < 0.05)
                        thrpt:  [-11.122% -8.5705% -5.1197%]
                        Performance has regressed.
Benchmarking wasm_call/io string_input_args_echo_ret/1: Collecting 10 samples in estimated 5.0001 s (910k iterations                                                                                                                    wasm_call/io string_input_args_echo_ret/1                        
                        time:   [5.1562 us 5.2643 us 5.5441 us]
                        thrpt:  [176.14 KiB/s 185.51 KiB/s 189.40 KiB/s]
                 change:
                        time:   [-12.534% -7.7261% -1.5664%] (p = 0.02 < 0.05)
                        thrpt:  [+1.5914% +8.3731% +14.330%]
                        Performance has improved.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) high mild
  1 (10.00%) high severe
Benchmarking wasm_call/io string_input_args_echo_ret/1000: Collecting 10 samples in estimated 5.0006 s (261k iterati                                                                                                                    wasm_call/io string_input_args_echo_ret/1000                        
                        time:   [18.930 us 18.994 us 19.064 us]
                        thrpt:  [50.024 MiB/s 50.210 MiB/s 50.379 MiB/s]
                 change:
                        time:   [-10.980% -8.8658% -7.1977%] (p = 0.00 < 0.05)
                        thrpt:  [+7.7559% +9.7283% +12.334%]
                        Performance has improved.
Benchmarking wasm_call/io string_input_args_echo_ret/1000000: Collecting 10 samples in estimated 5.1113 s (385 itera                                                                                                                    wasm_call/io string_input_args_echo_ret/1000000                        
                        time:   [13.280 ms 13.331 ms 13.408 ms]
                        thrpt:  [71.125 MiB/s 71.538 MiB/s 71.813 MiB/s]
                 change:
                        time:   [-13.808% -11.572% -9.1482%] (p = 0.00 < 0.05)
                        thrpt:  [+10.069% +13.086% +16.020%]
                        Performance has improved.
Benchmarking wasm_call/io string_input_args_empty_ret/0: Collecting 10 samples in estimated 5.0002 s (963k iteration                                                                                                                    wasm_call/io string_input_args_empty_ret/0                        
                        time:   [5.1895 us 5.2024 us 5.2142 us]
                        thrpt:  [0.0000   B/s 0.0000   B/s 0.0000   B/s]
                 change:
                        time:   [-11.208% -9.7535% -8.3752%] (p = 0.00 < 0.05)
                        thrpt:  [+9.1408% +10.808% +12.623%]
                        Performance has improved.
Benchmarking wasm_call/io string_input_args_empty_ret/1: Collecting 10 samples in estimated 5.0001 s (950k iteration                                                                                                                    wasm_call/io string_input_args_empty_ret/1                        
                        time:   [5.2647 us 5.2977 us 5.3420 us]
                        thrpt:  [182.81 KiB/s 184.34 KiB/s 185.49 KiB/s]
                 change:
                        time:   [-18.447% -15.511% -12.797%] (p = 0.00 < 0.05)
                        thrpt:  [+14.675% +18.359% +22.619%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
Benchmarking wasm_call/io string_input_args_empty_ret/1000: Collecting 10 samples in estimated 5.0002 s (451k iterat                                                                                                                    wasm_call/io string_input_args_empty_ret/1000                        
                        time:   [10.984 us 11.023 us 11.081 us]
                        thrpt:  [86.061 MiB/s 86.514 MiB/s 86.827 MiB/s]
                 change:
                        time:   [-17.726% -14.446% -11.866%] (p = 0.00 < 0.05)
                        thrpt:  [+13.464% +16.886% +21.545%]
                        Performance has improved.
Benchmarking wasm_call/io string_input_args_empty_ret/1000000: Collecting 10 samples in estimated 5.1130 s (990 iter                                                                                                                    wasm_call/io string_input_args_empty_ret/1000000                        
                        time:   [5.0882 ms 5.1055 ms 5.1279 ms]
                        thrpt:  [185.98 MiB/s 186.79 MiB/s 187.43 MiB/s]
                 change:
                        time:   [-14.747% -13.798% -12.931%] (p = 0.00 < 0.05)
                        thrpt:  [+14.851% +16.007% +17.298%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
Benchmarking wasm_call/io string_input_ignored_empty_ret/0: Collecting 10 samples in estimated 5.0002 s (939k iterat                                                                                                                    wasm_call/io string_input_ignored_empty_ret/0                        
                        time:   [6.7805 us 7.2552 us 7.5405 us]
                        thrpt:  [0.0000   B/s 0.0000   B/s 0.0000   B/s]
                 change:
                        time:   [-10.336% -4.9621% +1.0732%] (p = 0.14 > 0.05)
                        thrpt:  [-1.0619% +5.2212% +11.527%]
                        No change in performance detected.
Benchmarking wasm_call/io string_input_ignored_empty_ret/1: Collecting 10 samples in estimated 5.0004 s (591k iterat                                                                                                                    wasm_call/io string_input_ignored_empty_ret/1                        
                        time:   [9.2108 us 9.4789 us 9.6218 us]
                        thrpt:  [101.50 KiB/s 103.03 KiB/s 106.02 KiB/s]
                 change:
                        time:   [-6.1259% -0.4955% +4.9235%] (p = 0.86 > 0.05)
                        thrpt:  [-4.6925% +0.4980% +6.5256%]
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) low mild
Benchmarking wasm_call/io string_input_ignored_empty_ret/1000: Collecting 10 samples in estimated 5.0088 s (13k iter                                                                                                                    wasm_call/io string_input_ignored_empty_ret/1000                        
                        time:   [476.49 us 505.33 us 525.56 us]
                        thrpt:  [1.8146 MiB/s 1.8872 MiB/s 2.0015 MiB/s]
                 change:
                        time:   [-4.2385% +1.5367% +7.9797%] (p = 0.64 > 0.05)
                        thrpt:  [-7.3900% -1.5135% +4.4261%]
                        No change in performance detected.
Benchmarking wasm_call/io string_input_ignored_empty_ret/1000000: Collecting 10 samples in estimated 5.4380 s (110 i                                                                                                                    wasm_call/io string_input_ignored_empty_ret/1000000                        
                        time:   [76.153 ms 87.974 ms 94.949 ms]
                        thrpt:  [10.044 MiB/s 10.840 MiB/s 12.523 MiB/s]
                 change:
                        time:   [-21.872% -5.4523% +12.501%] (p = 0.57 > 0.05)
                        thrpt:  [-11.112% +5.7667% +27.995%]
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) low mild
Benchmarking wasm_call/io bytes_input_args_empty_ret/0: Collecting 10 samples in estimated 5.0000 s (868k iterations                                                                                                                    wasm_call/io bytes_input_args_empty_ret/0                        
                        time:   [5.6529 us 5.6797 us 5.7256 us]
                        thrpt:  [0.0000   B/s 0.0000   B/s 0.0000   B/s]
                 change:
                        time:   [-6.6814% -5.5794% -4.5187%] (p = 0.00 < 0.05)
                        thrpt:  [+4.7325% +5.9091% +7.1598%]
                        Performance has improved.
Benchmarking wasm_call/io bytes_input_args_empty_ret/1: Collecting 10 samples in estimated 5.0001 s (923k iterations                                                                                                                    wasm_call/io bytes_input_args_empty_ret/1                        
                        time:   [5.3402 us 5.3978 us 5.4441 us]
                        thrpt:  [179.38 KiB/s 180.92 KiB/s 182.87 KiB/s]
                 change:
                        time:   [-9.2841% -7.8983% -6.8139%] (p = 0.00 < 0.05)
                        thrpt:  [+7.3121% +8.5757% +10.234%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
Benchmarking wasm_call/io bytes_input_args_empty_ret/1000: Collecting 10 samples in estimated 5.0004 s (414k iterati                                                                                                                    wasm_call/io bytes_input_args_empty_ret/1000                        
                        time:   [11.467 us 11.741 us 12.000 us]
                        thrpt:  [79.473 MiB/s 81.228 MiB/s 83.165 MiB/s]
                 change:
                        time:   [-1.8216% +0.9203% +3.7327%] (p = 0.58 > 0.05)
                        thrpt:  [-3.5984% -0.9119% +1.8554%]
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
Benchmarking wasm_call/io bytes_input_args_empty_ret/1000000: Collecting 10 samples in estimated 5.1761 s (935 itera                                                                                                                    wasm_call/io bytes_input_args_empty_ret/1000000                        
                        time:   [5.2193 ms 5.2437 ms 5.2959 ms]
                        thrpt:  [180.08 MiB/s 181.87 MiB/s 182.72 MiB/s]
                 change:
                        time:   [-4.2421% -3.4195% -2.4654%] (p = 0.00 < 0.05)
                        thrpt:  [+2.5277% +3.5406% +4.4300%]
                        Performance has improved.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
Benchmarking wasm_call/io bytes_input_ignored_empty_ret/0: Collecting 10 samples in estimated 5.0008 s (279k iterati                                                                                                                    wasm_call/io bytes_input_ignored_empty_ret/0                        
                        time:   [32.451 us 33.541 us 34.433 us]
                        thrpt:  [0.0000   B/s 0.0000   B/s 0.0000   B/s]
                 change:
                        time:   [-1.1930% +9.0642% +21.990%] (p = 0.18 > 0.05)
                        thrpt:  [-18.026% -8.3109% +1.2074%]
                        No change in performance detected.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) low mild
  1 (10.00%) high severe
Benchmarking wasm_call/io bytes_input_ignored_empty_ret/1: Collecting 10 samples in estimated 5.0003 s (140k iterati                                                                                                                    wasm_call/io bytes_input_ignored_empty_ret/1                        
                        time:   [34.432 us 36.108 us 37.455 us]
                        thrpt:  [26.073 KiB/s 27.046 KiB/s 28.362 KiB/s]
                 change:
                        time:   [-10.099% +11.630% +48.263%] (p = 0.43 > 0.05)
                        thrpt:  [-32.552% -10.418% +11.234%]
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high severe
Benchmarking wasm_call/io bytes_input_ignored_empty_ret/1000: Collecting 10 samples in estimated 5.0769 s (2640 iter                                                                                                                    wasm_call/io bytes_input_ignored_empty_ret/1000                        
                        time:   [1.7797 ms 1.8794 ms 1.9520 ms]
                        thrpt:  [500.28 KiB/s 519.63 KiB/s 548.72 KiB/s]
                 change:
                        time:   [-32.100% -6.0360% +19.170%] (p = 0.69 > 0.05)
                        thrpt:  [-16.086% +6.4237% +47.275%]
                        No change in performance detected.
Found 2 outliers among 10 measurements (20.00%)
  1 (10.00%) low severe
  1 (10.00%) high severe
Benchmarking wasm_call/io bytes_input_ignored_empty_ret/1000000: Warming up for 3.0000 s
Warning: Unable to complete 10 samples in 5.0s. You may wish to increase target time to 7.2s.
Benchmarking wasm_call/io bytes_input_ignored_empty_ret/1000000: Collecting 10 samples in estimated 7.2469 s (55 ite                                                                                                                    wasm_call/io bytes_input_ignored_empty_ret/1000000                        
                        time:   [144.34 ms 156.88 ms 164.36 ms]
                        thrpt:  [5.8022 MiB/s 6.0790 MiB/s 6.6074 MiB/s]
                 change:
                        time:   [+3.7707% +12.814% +21.010%] (p = 0.01 < 0.05)
                        thrpt:  [-17.362% -11.358% -3.6337%]
                        Performance has regressed.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) low mild
wasm_call/io bytes_input_args_echo_ret/0                                                                            
                        time:   [5.5457 us 5.5881 us 5.6473 us]
                        thrpt:  [0.0000   B/s 0.0000   B/s 0.0000   B/s]
                 change:
                        time:   [-1.2510% +0.0930% +1.6481%] (p = 0.90 > 0.05)
                        thrpt:  [-1.6214% -0.0929% +1.2668%]
                        No change in performance detected.
Found 1 outliers among 10 measurements (10.00%)
  1 (10.00%) high mild
wasm_call/io bytes_input_args_echo_ret/1                                                                            
                        time:   [5.7955 us 5.8153 us 5.8509 us]
                        thrpt:  [166.91 KiB/s 167.93 KiB/s 168.50 KiB/s]
                 change:
                        time:   [-0.9181% +0.5793% +1.8894%] (p = 0.46 > 0.05)
                        thrpt:  [-1.8543% -0.5759% +0.9266%]
                        No change in performance detected.
…